backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
aiohttp
discord.py
openai
google-generativeai
anthropic
tiktoken
Pillow
tavily-python